[mod/#47] BaseDialogFragment 수정
